rename ChatItem
