+++ src/components/Dialogs/ChatItem/ChatItem.tsx
@@ -3,7 +3,7 @@ import { Comment } from 'components/UI/Comment';
 
 import { IProps } from './types';
 
-import styles from './Item.module.scss';
+import styles from './ChatItem.module.scss';
 
 const dateCompare = (date1: Date, date2: Date = new Date()): boolean => {
   return (
@@ -13,7 +13,7 @@ const dateCompare = (date1: Date, date2: Date = new Date()): boolean => {
   );
 };
 
-export const Item: FunctionComponent<IProps> = (props) => {
+export const ChatItem: FunctionComponent<IProps> = (props) => {
   const {
     chatType,
     ava,
