+++ src/components/Dialogs/ChatItem/index.tsx
@@ -1,0 +1,1 @@
+export * from './ChatItem.container';
