--- src/components/Dialogs/Item/index.tsx
@@ -1,1 +1,0 @@
-export * from './Item.container';
